Wide-reaching changes to beginning to bring things in line.
The API spec has gotten more strict for HarperDB, and the design of the API has been quite lax with allowing extra keys in the API responses.
As such, some requests simply fail. So we need a way to incrementally move away from the current model.

What I've done is change the `client.opRequest` to accept an Operation interface which gives a layer of separation between building the request structs & what is sent off to
HarperDB.
This should help with creating a more developer-friendly interface, while allowing us to easily transform the data into a valid payload for the database.

There are still a number of outstanding tasks to be done before this can be successfully merged:
- [ ] Fix up remaining tests
- [ ] Investigate why the licensing issue is coming up in the testing
- [ ] Write out the README into something a little more substantial